Moving _require_connection from storage.api to helpers.
This is to blob and bucket can use it without import cycles.
