+++ gcloud/storage/test_api.py
@@ -349,44 +349,6 @@ class Test__BucketIterator(unittest2.TestCase):
         self.assertEqual(bucket.name, BLOB_NAME)
 
 
-class Test__require_connection(unittest2.TestCase):
-
-    def _callFUT(self, connection=None):
-        from gcloud.storage.api import _require_connection
-        return _require_connection(connection=connection)
-
-    def _monkey(self, connection):
-        from gcloud.storage._testing import _monkey_defaults
-        return _monkey_defaults(connection=connection)
-
-    def test_implicit_unset(self):
-        with self._monkey(None):
-            with self.assertRaises(EnvironmentError):
-                self._callFUT()
-
-    def test_implicit_unset_w_existing_batch(self):
-        CONNECTION = object()
-        with self._monkey(None):
-            with _NoCommitBatch(connection=CONNECTION):
-                self.assertEqual(self._callFUT(), CONNECTION)
-
-    def test_implicit_unset_passed_explicitly(self):
-        CONNECTION = object()
-        with self._monkey(None):
-            self.assertTrue(self._callFUT(CONNECTION) is CONNECTION)
-
-    def test_implicit_set(self):
-        IMPLICIT_CONNECTION = object()
-        with self._monkey(IMPLICIT_CONNECTION):
-            self.assertTrue(self._callFUT() is IMPLICIT_CONNECTION)
-
-    def test_implicit_set_passed_explicitly(self):
-        IMPLICIT_CONNECTION = object()
-        CONNECTION = object()
-        with self._monkey(IMPLICIT_CONNECTION):
-            self.assertTrue(self._callFUT(CONNECTION) is CONNECTION)
-
-
 class Http(object):
 
     _called_with = None
@@ -399,18 +361,3 @@ class Http(object):
     def request(self, **kw):
         self._called_with = kw
         return self._response, self._content
-
-
-class _NoCommitBatch(object):
-
-    def __init__(self, connection):
-        self._connection = connection
-
-    def __enter__(self):
-        from gcloud.storage.batch import _BATCHES
-        _BATCHES.push(self._connection)
-        return self._connection
-
-    def __exit__(self, *args):
-        from gcloud.storage.batch import _BATCHES
-        _BATCHES.pop()
